backend/src/chatbot/requirements.txt
# chatbot/requirements.txt
# Install with: pip install -r chatbot/requirements.txt

groq>=0.9.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
fastapi>=0.110.0
pydantic>=2.0.0
python-dotenv>=1.0.0
